Basic argument parsing implemented
The script will now parse the inputs and separate out the various flags
and parameters. Unfortunately, the use of getopt/eval set means that by
the time they're done with the input arguments, there is really no way
to tell if the input file name contains spaces or if there are multiple
file names that have been supplied. All semblance of shell-specific
quoting is gone by then.

So, the script will just check if whatever remains is a valid file that
can be read. If not, throw an error.

The script is ready to be executable. The main function will just list
the arguments it was able to parse.