Two more ways for the program to flag an error
The first is when the step to make a back up copy of the original file
fails. It could fail, for example, if the user doesn't have write
permissions in the directory. In this case, the script would rather not
do anything.

The second way to fail is if the iconv command fails to complete the
conversion. In this case, the main file is still untouched and we've
even made a backup. So, no harm done if we exit.