Adjust delay and renumber entries
Both these operations are done by srttool. The tool doesn't support long
options. The -r flag turns on renumbering and the -d flag accepts a
delay in seconds. The output is written out to stdout.

In order to calculate the delay in seconds for srttool, I've used awk
instead of bc, as awk is more likely to be installed.