Implement stretching of time intervals between entries
The srttool from the transcode package only allows stretching up to 60
seconds over 3600 seconds. This is an unnecessary limitation. Instead,
using mkvmerge/mkvextract from the mkvtoolnix package, arbitrary stretch
factors can be used. In essence, we're encoding an mkv file with nothing
but a subtitle and extracting the subtitle out.